@@ -63,43 +65,38 @@
     "jsarch": "bin/jsarch.js"
   },
   "dependencies": {
-    "@babel/parser": "^7.18.3",
+    "@babel/parser": "^7.18.13",
     "ast-types": "^0.14.2",
-    "bluebird": "^3.7.2",
-    "commander": "^9.2.0",
+    "commander": "^9.4.0",
     "debug": "^4.3.4",
     "deep-extend": "^0.6.0",
     "glob": "^8.0.3",
-    "knifecycle": "^12.0.4",
-    "packagerc": "^1.1.0",
-    "pkg-dir": "^5.0.0",
-    "yerror": "^6.0.2"
+    "knifecycle": "^14.1.0",
+    "pkg-dir": "^7.0.0",
+    "rc": "^1.2.8",
+    "yerror": "^6.1.1"
   },
   "devDependencies": {
-    "@babel/cli": "^7.17.10",
-    "@babel/core": "^7.18.2",
-    "@babel/eslint-parser": "^7.18.2",
-    "@babel/plugin-proposal-object-rest-spread": "^7.18.0",
-    "@babel/preset-env": "^7.18.2",
-    "@babel/register": "^7.17.7",
-    "@mermaid-js/mermaid-cli": "^8.13.3",
-    "babel-plugin-knifecycle": "^5.0.2",
-    "commitizen": "^4.2.4",
+    "@mermaid-js/mermaid-cli": "^9.1.6",
+    "@typescript-eslint/eslint-plugin": "^5.36.0",
+    "@typescript-eslint/parser": "^5.36.0",
+    "commitizen": "^4.2.5",
     "conventional-changelog-cli": "^2.2.2",
     "coveralls": "^3.1.1",
     "cz-conventional-changelog": "^3.3.0",
-    "eslint": "^8.16.0",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-prettier": "^4.0.0",
-    "eslint-plugin-promise": "^4.3.1",
+    "esbuild": "^0.15.6",
+    "esbuild-jest": "^0.5.0",
+    "eslint": "^8.23.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-plugin-prettier": "^4.2.1",
+    "jest": "^29.0.1",
     "jsdoc-to-markdown": "^7.1.1",
-    "mermaid": "^8.13.3",
-    "metapak": "^4.0.3",
-    "metapak-nfroidure": "11.2.1",
-    "mocha": "^10.0.0",
-    "nyc": "^15.1.0",
-    "prettier": "^2.6.2",
-    "sinon": "^14.0.0"
+    "mermaid": "^9.1.6",
+    "metapak": "^4.0.4",
+    "metapak-nfroidure": "12.2.0",
+    "prettier": "^2.7.1",
+    "rimraf": "^3.0.2",
+    "typescript": "^4.8.2"
   },
   "author": {
     "name": "Nicolas Froidure",
@@ -112,32 +109,12 @@
   },
   "homepage": "https://github.com/nfroidure/jsarch#readme",
   "engines": {
-    "node": ">=12.19.0"
+    "node": ">=16.15.0"
   },
   "config": {
     "commitizen": {
       "path": "./node_modules/cz-conventional-changelog"
     }
-  },
-  "greenkeeper": {
-    "ignore": [
-      "commitizen",
-      "cz-conventional-changelog",
-      "conventional-changelog-cli",
-      "eslint",
-      "eslint-config-prettier",
-      "prettier",
-      "@babel/cli",
-      "@babel/core",
-      "@babel/register",
-      "@babel/preset-env",
-      "@babel/plugin-proposal-object-rest-spread",
-      "babel-eslint",
-      "mocha",
-      "coveralls",
-      "nyc",
-      "jsdoc-to-markdown"
-    ]
   },
   "contributors": [
     {
