@@ -1,8 +1,16 @@
-import assert from 'assert';
-import sinon from 'sinon';
-import Knifecycle, { constant } from 'knifecycle';
-import initJSArch, { DEFAULT_CONFIG } from './jsarch';
-import initParser from './parser';
+import {
+  describe,
+  beforeEach,
+  afterEach,
+  test,
+  expect,
+  jest,
+} from '@jest/globals';
+import { Knifecycle, constant } from 'knifecycle';
+import initJSArch, { DEFAULT_CONFIG } from './jsarch.js';
+import initParser from './parser.js';
+import type Glob from 'glob';
+import fs from 'fs';
 
 const JSARCH_PREFIX = `[//]: # ( )
 [//]: # (This file is automatically generated by the \`jsarch\`)
@@ -12,43 +20,39 @@ const JSARCH_PREFIX = `[//]: # ( )
 `;
 
 describe('jsArch service', () => {
-  const logStub = sinon.stub();
-  const globStub = sinon.stub();
-  const readFileAsyncStub = sinon.stub();
+  const log = jest.fn();
+  const glob = jest.fn<Glob>();
+  const readFile = jest.fn<typeof fs.promises.readFile>();
   let $;
 
   beforeEach(() => {
     $ = new Knifecycle();
     $.register(constant('EOL', '\n'));
     $.register(constant('CONFIG', DEFAULT_CONFIG));
-    $.register(constant('log', logStub));
-    $.register(constant('glob', globStub));
-    $.register(constant('fs', { readFileAsync: readFileAsyncStub }));
+    $.register(constant('log', log));
+    $.register(constant('glob', glob));
+    $.register(constant('fs', { readFile: readFile }));
     $.register(initParser);
     $.register(initJSArch);
   });
 
   afterEach(() => {
-    logStub.reset();
-    globStub.reset();
-    readFileAsyncStub.reset();
+    log.mockReset();
+    glob.mockReset();
+    readFile.mockReset();
   });
 
-  it('with no architecture notes', async () => {
-    globStub.returns(
-      Promise.resolve([
-        '/home/me/project/lulz.js',
-        '/home/me/project/kikoo.js',
-      ]),
-    );
+  test('with no architecture notes', async () => {
+    glob.mockResolvedValueOnce([
+      '/home/me/project/lulz.js',
+      '/home/me/project/kikoo.js',
+    ]);
 
-    readFileAsyncStub.returns(
-      Promise.resolve(`
+    readFile.mockResolvedValue(`
 
 console.log('test');
 
-    `),
-    );
+    `);
 
     const { jsArch } = await $.run(['jsArch']);
     const content = await jsArch({
@@ -56,18 +60,17 @@ console.log('test');
       base: './blob/master',
       cwd: '/home/me/project',
     });
-    assert.deepEqual(readFileAsyncStub.args, [
+    expect(readFile.mock.calls).toEqual([
       ['/home/me/project/lulz.js', 'utf-8'],
       ['/home/me/project/kikoo.js', 'utf-8'],
     ]);
-    assert.equal(content, '');
+    expect(content).toEqual('');
   });
 
-  it('with some architecture notes in a file', async () => {
-    globStub.returns(Promise.resolve(['/home/me/project/kikoo.js']));
+  test('with some architecture notes in a file', async () => {
+    glob.mockResolvedValueOnce(['/home/me/project/kikoo.js']);
 
-    readFileAsyncStub.returns(
-      Promise.resolve(`
+    readFile.mockResolvedValueOnce(`
 
 /* Architecture Note #1: Title
 
@@ -76,8 +79,7 @@ Some content !
 
 console.log('test');
 
-    `),
-    );
+    `);
 
     const { jsArch } = await $.run(['jsArch']);
     const content = await jsArch({
@@ -85,11 +87,10 @@ console.log('test');
       base: './blob/master',
       cwd: '/home/me/project',
     });
-    assert.deepEqual(readFileAsyncStub.args, [
+    expect(readFile.mock.calls).toEqual([
       ['/home/me/project/kikoo.js', 'utf-8'],
     ]);
-    assert.equal(
-      content,
+    expect(content).toEqual(
       `${JSARCH_PREFIX}# Architecture Notes
 
 ## Summary
@@ -107,11 +108,10 @@ Some content !
     );
   });
 
-  it('with some architecture notes containing a $', async () => {
-    globStub.returns(Promise.resolve(['/home/me/project/kikoo.js']));
+  test('with some architecture notes containing a $', async () => {
+    glob.mockResolvedValueOnce(['/home/me/project/kikoo.js']);
 
-    readFileAsyncStub.returns(
-      Promise.resolve(`
+    readFile.mockResolvedValueOnce(`
 
 /* Architecture Note #1: \`$autoload\`
 
@@ -120,8 +120,7 @@ Some content !
 
 console.log('test');
 
-    `),
-    );
+    `);
 
     const { jsArch } = await $.run(['jsArch']);
     const content = await jsArch({
@@ -129,11 +128,10 @@ console.log('test');
       base: './blob/master',
       cwd: '/home/me/project',
     });
-    assert.deepEqual(readFileAsyncStub.args, [
+    expect(readFile.mock.calls).toEqual([
       ['/home/me/project/kikoo.js', 'utf-8'],
     ]);
-    assert.equal(
-      content,
+    expect(content).toEqual(
       `${JSARCH_PREFIX}# Architecture Notes
 
 ## Summary
@@ -151,7 +149,7 @@ Some content !
     );
   });
 
-  it('with some architecture notes in a TypeScript file', async () => {
+  test('with some architecture notes in a TypeScript file', async () => {
     $.register(
       constant('CONFIG', {
         ...DEFAULT_CONFIG,
@@ -161,10 +159,9 @@ Some content !
         },
       }),
     );
-    globStub.returns(Promise.resolve(['/home/me/project/kikoo.js']));
+    glob.mockResolvedValueOnce(['/home/me/project/kikoo.js']);
 
-    readFileAsyncStub.returns(
-      Promise.resolve(`
+    readFile.mockResolvedValueOnce(`
 
 interface lol {
   lol : string;
@@ -177,8 +174,7 @@ Some content !
 
 console.log('test');
 
-    `),
-    );
+    `);
 
     const { jsArch } = await $.run(['jsArch']);
     const content = await jsArch({
@@ -186,11 +182,10 @@ console.log('test');
       base: './blob/master',
       cwd: '/home/me/project',
     });
-    assert.deepEqual(readFileAsyncStub.args, [
+    expect(readFile.mock.calls).toEqual([
       ['/home/me/project/kikoo.js', 'utf-8'],
     ]);
-    assert.equal(
-      content,
+    expect(content).toEqual(
       `${JSARCH_PREFIX}# Architecture Notes
 
 ## Summary
@@ -208,7 +203,7 @@ Some content !
     );
   });
 
-  it('with some architecture notes in a file and bitbucket links', async () => {
+  test('with some architecture notes in a file and bitbucket links', async () => {
     $.register(
       constant('CONFIG', {
         ...DEFAULT_CONFIG,
@@ -216,10 +211,9 @@ Some content !
       }),
     );
 
-    globStub.returns(Promise.resolve(['/home/me/project/kikoo.js']));
+    glob.mockResolvedValueOnce(['/home/me/project/kikoo.js']);
 
-    readFileAsyncStub.returns(
-      Promise.resolve(`
+    readFile.mockResolvedValueOnce(`
 
 /* Architecture Note #1: Title
 
@@ -228,8 +222,7 @@ Some content !
 
 console.log('test');
 
-    `),
-    );
+    `);
 
     const { jsArch } = await $.run(['jsArch']);
     const content = await jsArch({
@@ -237,11 +230,10 @@ console.log('test');
       base: './blob/master',
       cwd: '/home/me/project',
     });
-    assert.deepEqual(readFileAsyncStub.args, [
+    expect(readFile.mock.calls).toEqual([
       ['/home/me/project/kikoo.js', 'utf-8'],
     ]);
-    assert.equal(
-      content,
+    expect(content).toEqual(
       `${JSARCH_PREFIX}# Architecture Notes
 
 ## Summary
@@ -259,11 +251,10 @@ Some content !
     );
   });
 
-  it('with some indented architecture notes in a file', async () => {
-    globStub.returns(Promise.resolve(['/home/me/project/kikoo.js']));
+  test('with some indented architecture notes in a file', async () => {
+    glob.mockResolvedValueOnce(['/home/me/project/kikoo.js']);
 
-    readFileAsyncStub.returns(
-      Promise.resolve(`
+    readFile.mockResolvedValueOnce(`
 
     /* Architecture Note #1: Title
 
@@ -273,8 +264,7 @@ Some content !
 
 console.log('test');
 
-    `),
-    );
+    `);
 
     const { jsArch } = await $.run(['jsArch']);
     const content = await jsArch({
@@ -282,11 +272,10 @@ console.log('test');
       base: './blob/master',
       cwd: '/home/me/project',
     });
-    assert.deepEqual(readFileAsyncStub.args, [
+    expect(readFile.mock.calls).toEqual([
       ['/home/me/project/kikoo.js', 'utf-8'],
     ]);
-    assert.equal(
-      content,
+    expect(content).toEqual(
       `${JSARCH_PREFIX}# Architecture Notes
 
 ## Summary
@@ -305,16 +294,13 @@ Nice!
     );
   });
 
-  it('with architecture notes in several files', async () => {
-    globStub.returns(
-      Promise.resolve([
-        '/home/me/project/lulz.js',
-        '/home/me/project/kikoo.js',
-      ]),
-    );
+  test('with architecture notes in several files', async () => {
+    glob.mockResolvedValueOnce([
+      '/home/me/project/lulz.js',
+      '/home/me/project/kikoo.js',
+    ]);
 
-    readFileAsyncStub.onFirstCall().returns(
-      Promise.resolve(`
+    readFile.mockResolvedValueOnce(`
 /* Architecture Note #1.1: Title
 
 Some content !
@@ -326,11 +312,9 @@ console.log('test');
 Some content !
 */
 
-    `),
-    );
+    `);
 
-    readFileAsyncStub.onSecondCall().returns(
-      Promise.resolve(`
+    readFile.mockResolvedValueOnce(`
 /* Architecture Note #1.3: Title
 
 Some content !
@@ -342,8 +326,7 @@ console.log('test');
 Some content !
 */
 
-    `),
-    );
+    `);
 
     const { jsArch } = await $.run(['jsArch']);
     const content = await jsArch({
@@ -351,12 +334,11 @@ Some content !
       base: './blob/master',
       cwd: '/home/me/project',
     });
-    assert.deepEqual(readFileAsyncStub.args, [
+    expect(readFile.mock.calls).toEqual([
       ['/home/me/project/lulz.js', 'utf-8'],
       ['/home/me/project/kikoo.js', 'utf-8'],
     ]);
-    assert.equal(
-      content,
+    expect(content).toEqual(
       `${JSARCH_PREFIX}# Architecture Notes
 
 ## Summary
